src/problem5/crude/x/crude/types/genesis_test.go
package types_test

import (
	"testing"

	"crude/x/crude/types"

	"github.com/stretchr/testify/require"
)

func TestGenesisState_Validate(t *testing.T) {
	tests := []struct {
		desc     string
		genState *types.GenesisState
		valid    bool
	}{
		{
			desc:     "default is valid",
			genState: types.DefaultGenesis(),
			valid:    true,
		},
		{
			desc: "valid genesis state",
			genState: &types.GenesisState{

				ResourceList: []types.Resource{
					{
						Id: 0,
					},
					{
						Id: 1,
					},
				},
				ResourceCount: 2,
				// this line is used by starport scaffolding # types/genesis/validField
			},
			valid: true,
		},
		{
			desc: "duplicated resource",
			genState: &types.GenesisState{
				ResourceList: []types.Resource{
					{
						Id: 0,
					},
					{
						Id: 0,
					},
				},
			},
			valid: false,
		},
		{
			desc: "invalid resource count",
			genState: &types.GenesisState{
				ResourceList: []types.Resource{
					{
						Id: 1,
					},
				},
				ResourceCount: 0,
			},
			valid: false,
		},
		// this line is used by starport scaffolding # types/genesis/testcase
	}
	for _, tc := range tests {
		t.Run(tc.desc, func(t *testing.T) {
			err := tc.genState.Validate()
			if tc.valid {
				require.NoError(t, err)
			} else {
				require.Error(t, err)
			}
		})
	}
}
